Updated the Camera app
Unfortunately one of the images used for the Starship theme most likely a .9 image will break the camera if compiled from source. It however has no issue opening the apk as a 7zip archive and moving the images in after the apk has been built. So unfortunately until identifyingg the offending image or images,plurall before updating will need to first compile without the added Camera apk and then add it in aftertransferringg all the images. Arg!